requirements.txt
absl-py==1.1.0
alembic==1.14.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blessed==1.20.0
blinker==1.8.2
boltons @ file:///croot/boltons_1677628692245/work
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028895253/work
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1725551672989/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1625807838443/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
conda @ file:///croot/conda_1689269889729/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769286105/work
cycler==0.11.0
databricks-sdk==0.50.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
docker==7.1.0
einops==0.8.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.16.1
flask==3.0.3
flatbuffers==25.2.10
fonttools==4.33.3
fsspec==2025.3.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.8.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.46.3
gunicorn==23.0.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.35.1
importlib-metadata==8.5.0
importlib-resources==5.8.0
ipdb==0.13.13
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.8
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyter-server==1.17.1
jupyterlab==3.4.3
jupyterlab-language-pack-zh-CN==3.4.post1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.3
lazy-loader==0.4
lmdb==1.6.2
mako==1.3.10
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mlflow==2.17.2
mlflow-skinny==2.17.2
monai-weekly==1.4.dev2427
mpmath==1.3.0
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
nibabel==5.2.1
notebook==6.4.12
notebook-shim==0.1.0
numexpr @ file:///tmp/build/80754af9/numexpr_1640704208950/work
numpy==1.22.4
nvidia-ml-py==12.570.86
oauthlib==3.2.0
onnxruntime==1.19.2
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging @ file:///croot/packaging_1720101850331/work
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy @ file:///tmp/build/80754af9/pluggy_1648042571233/work
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.12.0
pynrrd==1.1.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz @ file:///croot/pytz_1713974312559/work
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==23.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.8
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
ruamel.yaml @ file:///tmp/build/80754af9/ruamel.yaml_1600552852248/work
ruamel.yaml.clib @ file:///opt/conda/conda-bld/ruamel.yaml.clib_1641552548372/work
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.2
sniffio==1.2.0
soupsieve==2.3.2.post1
sqlalchemy==2.0.40
sqlparse==0.5.3
stack-data==0.3.0
supervisor==4.2.4
sympy==1.13.3
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tinycss2==1.1.1
tokenizers==0.20.3
tomli==2.2.1
toolz @ file:///croot/toolz_1667464077321/work
torch @ http://download.pytorch.org/whl/cu113/torch-1.11.0%2Bcu113-cp38-cp38-linux_x86_64.whl
torchvision @ http://download.pytorch.org/whl/cu113/torchvision-0.12.0%2Bcu113-cp38-cp38-linux_x86_64.whl
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets==5.3.0
transformers==4.46.3
typing-extensions==4.13.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
werkzeug==3.0.6
widgetsnbextension==3.6.0
wrapt==1.17.2
zipp==3.20.2
